# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Free Win",
    "description": "\nFree Win centraliza los **Pedidos** de cartas de Yu-Gi-Oh! difíciles de conseguir\nen el país y las **Órdenes** que los usuarios envían dentro de cada Pedido.\n\nLa API permite gestionar usuarios y permisos, abrir y administrar Pedidos, y\nrevisar las Órdenes que referencian publicaciones del servicio de búsqueda.\n\n## Autenticación temporal\n\nDurante el desarrollo local, la identidad se selecciona mediante la configuración\ndel backend. Los clientes no envían todavía credenciales, tokens Bearer ni OAuth2.\nEste mecanismo es temporal y por ello OpenAPI no declara un esquema de seguridad.\n",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/order-requests/": {
      "post": {
        "tags": [
          "order-requests"
        ],
        "summary": "Enviar una Orden",
        "description": "Crea una Orden propia dentro de un Pedido abierto. Todas las publicaciones de carta deben existir antes de guardar la Orden.",
        "operationId": "createOrderRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "No existe una identidad autenticada válida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "La identidad no posee el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "El Pedido o una publicación de carta no existe.",
            "content": {
              "application/json": {
                "examples": {
                  "period": {
                    "summary": "Pedido inexistente",
                    "value": {
                      "detail": "El Pedido no existe"
                    }
                  },
                  "listing": {
                    "summary": "Publicación inexistente",
                    "value": {
                      "detail": "La publicación de carta no existe"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "El Pedido no está abierto para recibir Órdenes.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "El Pedido no está abierto para recibir Órdenes"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "La entrada no cumple el contrato de la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "order-requests"
        ],
        "summary": "Listar Órdenes visibles",
        "description": "Devuelve únicamente las Órdenes propias para un Usuario convencional. Una identidad con permiso global puede consultar todas las Órdenes.",
        "operationId": "listOrderRequests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Página solicitada; la primera página es 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Página solicitada; la primera página es 1."
          },
          {
            "name": "shows",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Cantidad máxima de Órdenes por página.",
              "default": 100,
              "title": "Shows"
            },
            "description": "Cantidad máxima de Órdenes por página."
          },
          {
            "name": "orderPeriodId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limita el listado a un Pedido concreto.",
              "title": "Orderperiodid"
            },
            "description": "Limita el listado a un Pedido concreto."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderRequestStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limita el listado al estado indicado.",
              "title": "Status"
            },
            "description": "Limita el listado al estado indicado."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRequestListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No existe una identidad autenticada válida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "La identidad no posee el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "La entrada no cumple el contrato de la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/order-requests/{order_request_id}": {
      "patch": {
        "tags": [
          "order-requests"
        ],
        "summary": "Actualizar la nota compartida de una Orden",
        "description": "Sustituye o elimina la nota compartida. La Orden se bloquea durante la operación y solo se auditan cambios efectivos.",
        "operationId": "updateOrderRequestNote",
        "parameters": [
          {
            "name": "order_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Identificador positivo de la Orden.",
              "title": "Order Request Id"
            },
            "description": "Identificador positivo de la Orden."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "No existe una identidad autenticada válida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "La identidad no posee el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "La Orden no existe o pertenece a otro Usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "La transición o sus precondiciones no son válidas.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "La entrada no cumple el contrato de la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "order-requests"
        ],
        "summary": "Consultar una Orden",
        "description": "Devuelve una Orden propia o cualquier Orden cuando la identidad posee permiso global. Una Orden ajena se presenta como inexistente.",
        "operationId": "getOrderRequest",
        "parameters": [
          {
            "name": "order_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Identificador positivo de la Orden.",
              "title": "Order Request Id"
            },
            "description": "Identificador positivo de la Orden."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRequestResponse"
                }
              }
            }
          },
          "401": {
            "description": "No existe una identidad autenticada válida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "La identidad no posee el permiso requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "La Orden no existe o pertenece a otro Usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "La entrada no cumple el contrato de la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Mensaje comprensible que explica el error recuperable.",
            "examples": [
              "El recurso no existe"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "OrderRequestCreate": {
        "properties": {
          "orderPeriodId": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Orderperiodid",
            "description": "Pedido abierto dentro del cual se envía la Orden.",
            "examples": [
              12
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Nota opcional compartida entre el usuario y los administradores.",
            "examples": [
              "Priorizar cartas en condición Near Mint."
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderRequestItemCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items",
            "description": "Cartas solicitadas; una publicación no puede repetirse."
          }
        },
        "type": "object",
        "required": [
          "orderPeriodId",
          "items"
        ],
        "title": "OrderRequestCreate"
      },
...
      "OrderRequestListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderRequestResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "OrderRequestListResponse"
      },
...
      "OrderRequestStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "in_review",
          "accepted",
          "rejected",
          "cancelled"
        ],
        "title": "OrderRequestStatus"
      },
...
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse"
      }
    }
  },
...
}